Split logic for making logs into its own class
