+++ Source/AssetGenerator.csproj
@@ -55,6 +55,7 @@
     <Compile Include="Common.cs" />
     <Compile Include="Data.cs" />
     <Compile Include="Enums.cs" />
+    <Compile Include="Logs.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Runtime\GLTF.cs" />
