+++ Source/Logs.cs
@@ -1,0 +1,192 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.IO;
+
+namespace AssetGenerator
+{
+    class LogBuilder
+    {
+        StringBuilder csv = new StringBuilder();
+        StringBuilder md = new StringBuilder();
+        List<string> mdLogHeader = new List<string>();
+        List<List<string>> mdLogPrereqs = new List<List<string>>();
+        List<List<string>> mdLog = new List<List<string>>();
+        string lastName = null;
+
+        public LogBuilder()
+        {
+
+        }
+        
+        public void SetupHeader(TestValues test)
+        {
+            // Setup the log file header
+            mdLogHeader.Add("The following table shows the properties that are set for every model.  ");
+            mdLogHeader.Add("No attributes are explicitly set in every model generated by this test, other than those in the base model.  ");
+            mdLogHeader.Add("The following table shows the properties that are set for a given model.  ");
+            mdLogHeader.Add("All values of Byte and Short are normalized unsigned.  ");
+
+            if (test.requiredAttributes != null)
+            {
+                // List attributes that are set in every generated model (prerequisites)
+                mdLogPrereqs.Add(new List<string>()); // First line of table must be blank
+                mdLogPrereqs.Add(new List<string>
+                    {
+                    "Attribute", // First cells are a static label
+                    "**Values**"
+                    });
+                mdLogPrereqs.Add(new List<string>
+                    {
+                    ":---:", // Hyphens for row after header
+                    ":---:",
+                    });
+                for (int i = 0; i < test.requiredAttributes.Count; i++)
+                {
+                    string attributeName;
+                    attributeName = test.requiredAttributes[i].name.ToString();
+                    attributeName = LogStringHelper.GenerateNameWithSpaces(attributeName);
+                    mdLogPrereqs.Add(new List<string>
+                        {
+                        attributeName,
+                        LogStringHelper.ConvertTestValueToString(test.requiredAttributes[i])
+                        });
+                }
+            }
+
+            // Now start the table for generated models
+            mdLog.Add(new List<string>()); // First line of table must be blank
+            mdLog.Add(new List<string>
+                {
+                    "Index" // First cell is a static header name
+                });
+            mdLog.Add(new List<string>
+                {
+                    ":---:" // Hyphens for row after header 
+                });
+            for (int i = 0; i < test.attributes.Count; i++)
+            {
+                string attributeName;
+                if (test.attributes[i].prerequisite != AttributeName.Undefined && test.attributes[i].attributeGroup == 0)
+                {
+                    attributeName = test.attributes[i].prerequisite.ToString() + test.attributes[i].name.ToString();
+                }
+                else
+                {
+                    attributeName = test.attributes[i].name.ToString();
+                }
+                attributeName = LogStringHelper.GenerateNameWithSpaces(attributeName);
+                if (attributeName != lastName) // Skip duplicate names caused by non-binary attributes
+                {
+                    lastName = attributeName;
+                    mdLog[1].Add(attributeName);
+                    mdLog[2].Add(":---:");
+                }
+            }
+        }
+
+        public void SetupTable(TestValues test, int comboIndex, List<List<Attribute>> combos)
+        {
+            mdLog.Add(new List<string> // New row for a new model
+                    {
+                        // Displays the number of the model and is a link to the model
+                        '[' + comboIndex.ToString() + "](./" + test.testType.ToString() + '_' + comboIndex + ".gltf)"
+                    });
+            int logIndex = mdLog.Count - 1;
+            List<int> nonBinaryUsed = new List<int>();
+            foreach (var possibleAttribute in test.attributes)
+            {
+                var attributeIndex = combos[comboIndex].FindIndex(e =>
+                    e.name == possibleAttribute.name &&
+                    e.prerequisite == possibleAttribute.prerequisite);
+                if (attributeIndex != -1)
+                {
+                    if (possibleAttribute.attributeGroup > 0)
+                    {
+                        var alreadyUsed = nonBinaryUsed.Exists(x => x == possibleAttribute.attributeGroup);
+                        if (alreadyUsed)
+                        {
+                            // Overwrites the empty cell if a nonbinary of the same time had already been encountered and not used
+                            mdLog[logIndex][mdLog[logIndex].Count - 1] = LogStringHelper.ConvertTestValueToString(possibleAttribute);
+                        }
+                        else
+                        {
+                            // Creates a new cell, since this nonbinary type had not been encountered before
+                            mdLog[logIndex].Add(LogStringHelper.ConvertTestValueToString(possibleAttribute));
+                            nonBinaryUsed.Add(possibleAttribute.attributeGroup);
+                        }
+                    }
+                    else
+                    {
+                        mdLog[logIndex].Add(LogStringHelper.ConvertTestValueToString(possibleAttribute));
+                    }
+                }
+                else
+                {
+                    if (possibleAttribute.attributeGroup > 0)
+                    {
+                        var alreadyUsed = nonBinaryUsed.Exists(x => x == possibleAttribute.attributeGroup);
+                        if (!alreadyUsed)
+                        {
+                            mdLog[logIndex].Add(" ");
+                            nonBinaryUsed.Add(possibleAttribute.attributeGroup);
+                        }
+                    }
+                    else
+                    {
+                        mdLog[logIndex].Add(" ");
+                    }
+                }
+            }
+
+            string[] name = LogStringHelper.GenerateName(combos[comboIndex]);
+            var writeToLog = string.Format("{0},{1}", comboIndex, String.Join(" - ", name));
+            csv.AppendLine(writeToLog);
+        }
+
+        public void WriteOut(TestValues test, string assetFolder)
+        {
+            if (test.requiredAttributes != null)
+            {
+                md.AppendLine(mdLogHeader[0]); // Header for the prerequisite table. Only show if the table has values.
+                foreach (var line in mdLogPrereqs)
+                {
+                    md.AppendLine(String.Join(" | ", line));
+                }
+            }
+            else
+            {
+                md.AppendLine(mdLogHeader[1]); // If there are no prerequisites, this is the header.
+            }
+
+            md.AppendLine(" "); // Linebreak
+            md.AppendLine(mdLogHeader[2]); // Header for test grid table
+            foreach (var x in mdLog)
+            {
+                var resultByte = x.FirstOrDefault(s => s.Contains("Byte"));
+                if (resultByte != null)
+                {
+                    md.AppendLine(mdLogHeader[3]);
+                    break;
+                }
+                var resultShort = mdLog.FirstOrDefault(s => s.Contains("Short"));
+                if (resultShort != null)
+                {
+                    md.AppendLine(mdLogHeader[3]);
+                    break;
+                }
+            }
+
+            foreach (var line in mdLog)
+            {
+                md.AppendLine(String.Join(" | ", line));
+            }
+
+            var logFile = Path.Combine(assetFolder, test.testType.ToString() + "_log.csv");
+            File.WriteAllText(logFile, csv.ToString());
+            var mdLogFile = Path.Combine(assetFolder, "README.md");
+            File.WriteAllText(mdLogFile, md.ToString());
+        }
+    }
+}
