+++ Source/Program.cs
@@ -37,12 +37,13 @@ namespace AssetGenerator
                 TestValues makeTest = new TestValues();
                 makeTest.InitializeTestValues(test);
                 var combos = ComboHelper.AttributeCombos(makeTest);
-                var csv = new StringBuilder();
-                var md = new StringBuilder();
-                List<string> mdLogHeader = new List<string>(); ;
-                List<List<string>> mdLogPrereqs = new List<List<string>>();
-                List<List<string>> mdLog = new List<List<string>>();
-                string lastName = null;
+                //var csv = new StringBuilder();
+                //var md = new StringBuilder();
+                //List<string> mdLogHeader = new List<string>(); ;
+                //List<List<string>> mdLogPrereqs = new List<List<string>>();
+                //List<List<string>> mdLog = new List<List<string>>();
+                //string lastName = null;
+                LogBuilder logs = new LogBuilder();
 
                 // Delete any preexisting files in the output directories, then create those directories if needed
                 var assetFolder = Path.Combine(executingAssemblyFolder, test.ToString());
@@ -73,68 +74,7 @@ namespace AssetGenerator
 
                 Directory.CreateDirectory(assetFolder);
 
-                // Setup the log file header
-                mdLogHeader.Add("The following table shows the properties that are set for every model.  ");
-                mdLogHeader.Add("No attributes are explicitly set in every model generated by this test, other than those in the base model.  ");
-                mdLogHeader.Add("The following table shows the properties that are set for a given model.  ");
-                mdLogHeader.Add("All values of Byte and Short are normalized unsigned.  ");
-
-                if (makeTest.requiredAttributes != null)
-                {
-                    // List attributes that are set in every generated model (prerequisites)
-                    mdLogPrereqs.Add(new List<string>()); // First line of table must be blank
-                    mdLogPrereqs.Add(new List<string>
-                    {
-                    "Attribute", // First cells are a static label
-                    "**Values**"
-                    });
-                    mdLogPrereqs.Add(new List<string>
-                    {
-                    ":---:", // Hyphens for row after header
-                    ":---:",
-                    });
-                    for (int i = 0; i < makeTest.requiredAttributes.Count; i++)
-                    {
-                        string attributeName;
-                        attributeName = makeTest.requiredAttributes[i].name.ToString();
-                        attributeName = LogStringHelper.GenerateNameWithSpaces(attributeName);
-                        mdLogPrereqs.Add(new List<string>
-                        {
-                        attributeName,
-                        LogStringHelper.ConvertTestValueToString(makeTest.requiredAttributes[i])
-                        });
-                    }
-                }
-
-                // Now start the table for generated models
-                mdLog.Add(new List<string>()); // First line of table must be blank
-                mdLog.Add(new List<string>
-                {
-                    "Index" // First cell is a static header name
-                });
-                mdLog.Add(new List<string>
-                {
-                    ":---:" // Hyphens for row after header 
-                });
-                for (int i = 0; i < makeTest.attributes.Count; i++)
-                {
-                    string attributeName;
-                    if (makeTest.attributes[i].prerequisite != AttributeName.Undefined && makeTest.attributes[i].attributeGroup == 0)
-                    {
-                        attributeName = makeTest.attributes[i].prerequisite.ToString() + makeTest.attributes[i].name.ToString();
-                    }
-                    else
-                    {
-                        attributeName =  makeTest.attributes[i].name.ToString();
-                    }
-                    attributeName = LogStringHelper.GenerateNameWithSpaces(attributeName);
-                    if (attributeName != lastName) // Skip duplicate names caused by non-binary attributes
-                    {
-                        lastName = attributeName;
-                        mdLog[1].Add(attributeName);
-                        mdLog[2].Add(":---:");
-                    }
-                }
+                logs.SetupHeader(makeTest);
 
                 int numCombos = combos.Count;
                 for (int comboIndex = 0; comboIndex < numCombos; comboIndex++)
@@ -464,102 +404,10 @@ namespace AssetGenerator
                         File.WriteAllBytes(dataFile, ((MemoryStream)data.Writer.BaseStream).ToArray());
                     }
 
-                    mdLog.Add(new List<string> // New row for a new model
-                    {
-                        // Displays the number of the model and is a link to the model
-                        '[' + comboIndex.ToString() + "](./" + test.ToString() + '_' + comboIndex + ".gltf)"
-                    }); 
-                    int logIndex = mdLog.Count - 1;
-                    List<int> nonBinaryUsed = new List<int>();
-                    foreach (var possibleAttribute in makeTest.attributes)
-                    {
-                        var attributeIndex = combos[comboIndex].FindIndex(e =>
-                            e.name == possibleAttribute.name &&
-                            e.prerequisite == possibleAttribute.prerequisite);
-                        if (attributeIndex != -1)
-                        {
-                            if (possibleAttribute.attributeGroup > 0)
-                            {
-                                var alreadyUsed = nonBinaryUsed.Exists(x => x == possibleAttribute.attributeGroup);
-                                if (alreadyUsed)
-                                {
-                                    // Overwrites the empty cell if a nonbinary of the same time had already been encountered and not used
-                                    mdLog[logIndex][mdLog[logIndex].Count - 1] = LogStringHelper.ConvertTestValueToString(possibleAttribute);
-                                }
-                                else
-                                {
-                                    // Creates a new cell, since this nonbinary type had not been encountered before
-                                    mdLog[logIndex].Add(LogStringHelper.ConvertTestValueToString(possibleAttribute));
-                                    nonBinaryUsed.Add(possibleAttribute.attributeGroup);
-                                }
-                            }
-                            else
-                            {
-                                mdLog[logIndex].Add(LogStringHelper.ConvertTestValueToString(possibleAttribute));
-                            }
-                        }
-                        else
-                        {
-                            if (possibleAttribute.attributeGroup > 0)
-                            {
-                                var alreadyUsed = nonBinaryUsed.Exists(x => x == possibleAttribute.attributeGroup);
-                                if (!alreadyUsed)
-                                {
-                                    mdLog[logIndex].Add(" ");
-                                    nonBinaryUsed.Add(possibleAttribute.attributeGroup);
-                                }
-                            }
-                            else
-                            {
-                                mdLog[logIndex].Add(" ");
-                            }
-                        }
-                    }
-
-                    var writeToLog = string.Format("{0},{1}", comboIndex, String.Join(" - ", name));
-                    csv.AppendLine(writeToLog);
+                    logs.SetupTable(makeTest,comboIndex, combos);
                 }
 
-                if (makeTest.requiredAttributes != null)
-                {
-                    md.AppendLine(mdLogHeader[0]); // Header for the prerequisite table. Only show if the table has values.
-                    foreach (var line in mdLogPrereqs)
-                    {
-                        md.AppendLine(String.Join(" | ", line));
-                    }
-                }
-                else
-                {
-                    md.AppendLine(mdLogHeader[1]); // If there are no prerequisites, this is the header.
-                }
-
-                md.AppendLine(" "); // Linebreak
-                md.AppendLine(mdLogHeader[2]); // Header for test grid table
-                foreach (var x in mdLog)
-                {
-                    var resultByte = x.FirstOrDefault(s => s.Contains("Byte"));
-                    if (resultByte != null)
-                    {
-                        md.AppendLine(mdLogHeader[3]);
-                        break;
-                    }
-                    var resultShort = mdLog.FirstOrDefault(s => s.Contains("Short"));
-                    if (resultShort != null)
-                    {
-                        md.AppendLine(mdLogHeader[3]);
-                        break;
-                    }
-                }
-
-                foreach (var line in mdLog)
-                {
-                    md.AppendLine(String.Join(" | ", line));
-                }
-                
-                var logFile = Path.Combine(assetFolder, test.ToString() + "_log.csv");
-                File.WriteAllText(logFile, csv.ToString());
-                var mdLogFile = Path.Combine(assetFolder, "README.md");
-                File.WriteAllText(mdLogFile, md.ToString());
+                logs.WriteOut(makeTest, assetFolder);
             }
             Console.WriteLine("Model Creation Complete!");
             Console.WriteLine("Completed in : " + TimeSpan.FromTicks(Stopwatch.GetTimestamp()).ToString());
